Moved all staking queries to a new Queries directory, into individual files
