+++ src/views/Staking/layout.tsx
@@ -20,12 +20,10 @@ import { useDisclosure } from "@chakra-ui/hooks";
 import ColoredText from "../../components/ColoredText";
 import { BigNumber, BigNumberish, constants, FixedNumber } from "ethers";
 import coloredAgaveLogo from "../../assets/image/colored-agave-logo.svg";
-import {
-  StakingCooldownInfo,
-  useStakingAgavePrice,
-  useTotalStakedForAllUsers,
-} from "./queries";
 import { useAppWeb3 } from "../../hooks/appWeb3";
+import { useStakingAgavePrice } from "../../queries/stakingAgavePrice";
+import { StakingCooldownInfo } from "../../queries/stakingCooldown";
+import { useTotalStakedForAllUsers } from "../../queries/totalStakedForAllUsers";
 
 export interface StakingBannerProps {}
